+++ feature/mypage/src/main/java/com/napzak/market/mypage/mypage/MyPageScreen.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.napzak.market.common.state.UiState
-import com.napzak.market.designsystem.component.NapzakLoadingOverlay
+import com.napzak.market.designsystem.component.loading.NapzakLoadingOverlay
 import com.napzak.market.designsystem.component.topbar.NapzakLogoTopBar
 import com.napzak.market.designsystem.theme.NapzakMarketTheme
 import com.napzak.market.mypage.mypage.component.MyPageMenuCard
